backend/requirements.txt
fastapi
chromadb
pydantic
langchain
langchain-community
langchain-gigachat
python-dotenv
tiktoken
python-docx
pandas
openpyxl
PyMuPDF
python-multipart
